+++ CHANGELOG.md
@@ -4,6 +4,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 ## Next Release (replace with git tag when deployed)
  * Bump runtimeVersion to `2025.01.15`.
  * Note: started deleting `Package.isBlocked`, `Publisher.isBlocked`, `User.isBlocked`.
+ * Note: Updates search form normalization: whitespaces.
 
 ## `20250114t095800-all`
  * Bump runtimeVersion to `2025.01.14`.